Medium 4 ms (52.19%) 42.2 MB (7.78%)

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.java
@@ -1,7 +1,7 @@
 class Solution {
     public boolean judgeSquareSum(int c) {
         long low = 0;
-        long high = (long) Math.sqrt(c);
+        long high = (int) Math.sqrt(c);
 
         while (low <= high) {
             long sum = low * low + high * high;
